drop_unused_value_labels {labelled}	R Documentation

Drop unused value labels

Description

Drop value labels associated to a value not present in the data.

Usage

drop_unused_value_labels(x)

Arguments

x
A vector, a data frame or a survey design.

Examples

x <- labelled(c(1, 2, 2, 1), c(yes = 1, no = 2, maybe = 3))
x
drop_unused_value_labels(x)

[Package labelled version 2.16.0 Index]