mean.haven_labelled {labelled}	R Documentation

Arithmetic mean for numeric labelled vectors

Description

Arithmetic mean for numeric labelled vectors

Usage

## S3 method for class 'haven_labelled'
mean(x, trim = 0, na.rm = FALSE, ...)

Arguments

x
a numeric labelled vector.

trim
the fraction (0 to 0.5) of observations to be trimmed from each end of x . before the mean is computed. Values of trim outside that range are taken as the nearest endpoint.

na.rm
a logical value indicating whether NA values should be stripped before the computation proceeds.

...
additional arguments to be passed to or from methods.

See Also

mean

[Package labelled version 2.0.2 Index]