names_prefixed_by_values {labelled}	R Documentation

Turn a named vector into a vector of names prefixed by values

Description

Turn a named vector into a vector of names prefixed by values

Usage

names_prefixed_by_values(x, sep = "[]")

Arguments

x
vector to be prefixed

sep
(string) separator between value and name

Examples

df <- dplyr::tibble(
  c1 = labelled(c("M", "M", "F"), c(Male = "M", Female = "F")),
  c2 = labelled(c(1, 1, 2), c(Yes = 1, No = 2))
)
val_labels(df$c1)
val_labels(df$c1) %>% names_prefixed_by_values()
val_labels(df)
val_labels(df) %>% names_prefixed_by_values()
val_labels(df) %>% names_prefixed_by_values(sep = ":")

[Package labelled version 2.16.0 Index]