update_labelled {labelled}	R Documentation

Update labelled data to last version

Description

Labelled data imported with haven version 1.1.2 or before or created with haven::labelled() version 1.1.0 or before was using "labelled" and "labelled_spss" classes.

Usage

update_labelled(x)

## S3 method for class 'labelled'
update_labelled(x)

## S3 method for class 'haven_labelled_spss'
update_labelled(x)

## S3 method for class 'haven_labelled'
update_labelled(x)

## S3 method for class 'data.frame'
update_labelled(x)

Arguments

x
An object (vector or data.frame) to convert.

Details

Since version 2.0.0 of these two packages, "haven_labelled" and "haven_labelled_spss" are used instead.

Since haven 2.3.0, "haven_labelled" class has been evolving using now vctrs package.

update_labelled() convert labelled vectors from the old to the new classes and to reconstruct all labelled vectors with the last version of the package.

See Also

haven::labelled(), haven::labelled_spss()

[Package labelled version 2.16.0 Index]