Look for keywords variable names and descriptions

look_for emulates the lookfor Stata command in R. It supports searching into the variable names of regular R data frames as well as into variable labels descriptions. The command is meant to help users finding variables in large datasets.

look_for(data, ..., labels = TRUE, ignore.case = TRUE, details = FALSE)

lookfor(data, ..., labels = TRUE, ignore.case = TRUE, details = FALSE)

Arguments

data
a data frame

...
list of keywords, a character string (or several character strings), which can be formatted as a regular expression suitable for a grep pattern, or a vector of keywords; displays all variables if not specified

labels
whether or not to search variable labels (descriptions); TRUE by default

ignore.case
whether or not to make the keywords case sensitive; TRUE by default (case is ignored during matching)

details
add details about each variable (see examples)

Source

Based on the behaviour of the lookfor command in Stata.

Value

a data frame featuring the variable position, name and description (if it exists) in the original data frame

Details

The function looks into the variable names for matches to the keywords. If available, variable labels are included in the search scope. Variable labels of data.frame imported with foreign or memisc packages will also be taken into account (see to_labelled).

look_for and lookfor are equivalent.

Examples

look_for(iris)

#>       variable
#> 1 Sepal.Length
#> 2  Sepal.Width
#> 3 Petal.Length
#> 4  Petal.Width
#> 5      Species

# Look for a single keyword.
look_for(iris, "petal")

#>       variable
#> 3 Petal.Length
#> 4  Petal.Width

look_for(iris, "s")

#>       variable
#> 1 Sepal.Length
#> 2  Sepal.Width
#> 5      Species

# Look for with a regular expression
look_for(iris, "petal|species")

#>       variable
#> 3 Petal.Length
#> 4  Petal.Width
#> 5      Species

look_for(iris, "s$")

#>   variable
#> 5  Species

# Look for with several keywords
look_for(iris, "pet", "sp")

#>       variable
#> 3 Petal.Length
#> 4  Petal.Width
#> 5      Species

look_for(iris, "pet", "sp", "width")

#>       variable
#> 2  Sepal.Width
#> 3 Petal.Length
#> 4  Petal.Width
#> 5      Species

# Labelled data
if (FALSE) require(questionr)
data(fertility)

#> Warning: data set 'fertility' not found

look_for(women)

#>   variable
#> 1   height
#> 2   weight

look_for(women, "date")

#> Nothing found. Sorry.

# Display details
look_for(women, details = TRUE)

#>   variable   class   type levels value_labels unique_values n_na na_values
#> 1   height numeric double                                15    0          
#> 2   weight numeric double                                15    0          
#>   na_range
#> 1         
#> 2