na_values {labelled}	R Documentation

Get / Set SPSS missing values

Description

Get / Set SPSS missing values

Usage

na_values(x)

na_values(x) <- value

na_range(x)

na_range(x) <- value

set_na_values(.data, ...)

set_na_range(.data, ...)

user_na_to_na(x)

Arguments

x
A vector.

value
A vector of values that should also be considered as missing (for na_values) or a numeric vector of length two giving the (inclusive) extents of the range (for na_values, use -Inf and Inf if you want the range to be open ended).

.data
a data frame

...
name-value pairs of missing values (see examples)

Details

See labelled_spss for a presentation of SPSS's user defined missing values. Note that it is mandatory to define value labels before defining missing values. You can use user_na_to_na to convert user defined missing values to NA.

Value

na_values will return a vector of values that should also be considered as missing. na_range will return a numeric vector of length two giving the (inclusive) extents of the range.

set_na_values and set_na_range will return an updated copy of .data.

Note

set_na_values and set_na_range could be used with dplyr.

See Also

labelled_spss, user_na_to_na

Examples

v <- labelled(c(1,2,2,2,3,9,1,3,2,NA), c(yes = 1, no = 3, "don't know" = 9))
v
na_values(v) <- 9
na_values(v)
v
user_na_to_na(v)
na_values(v) <- NULL
v
na_range(v) <- c(5, Inf)
na_range(v)
v
user_na_to_na(v)
if (require(dplyr)) {
  # setting value labels
  df <- data_frame(s1 = c("M", "M", "F", "F"), s2 = c(1, 1, 2, 9)) %>%
    set_value_labels(s2 = c(yes = 1, no = 2)) %>%
    set_na_values(s2 = 9)
  na_values(df)

  # removing missing values
  df <- df %>% set_na_values(s2 = NULL)
  df$s2
}

[Package labelled version 2.0.2 Index]