Remove variable label, value labels and user defined missing values

Use remove_var_label to remove variable label, remove_val_labels to remove value labels, remove_user_na to remove user defined missing values (na_values and na_range) and remove_labels to remove all.

remove_labels(x, user_na_to_na = FALSE)

remove_var_label(x)

remove_val_labels(x)

remove_user_na(x, user_na_to_na = FALSE)

user_na_to_na(x)

Arguments

x
A vector or a data frame.

user_na_to_na
Convert user defined missing values into NA?

Details

Be careful with remove_user_na and remove_labels, user defined missing values will not be automatically converted to NA, except if you specify user_na_to_na = TRUE. user_na_to_na(x) is an equivalent of remove_user_na(x, user_na_to_na = TRUE).

Examples

x1 <- labelled_spss(1:10, c(Good = 1, Bad = 8), na_values = c(9, 10))
var_label(x1) <- "A variable"
x1

#> <Labelled SPSS integer>
#>  [1]  1  2  3  4  5  6  7  8  9 10
#> Missing values: 9, 10
#> 
#> Labels:
#>  value label
#>      1  Good
#>      8   Bad

x2 <- remove_labels(x1)

#> Some user defined missing values have been removed but not converted to NA.

x2

#>  [1]  1  2  3  4  5  6  7  8  9 10

x3 <- remove_labels(x1, user_na_to_na = TRUE)
x3

#>  [1]  1  2  3  4  5  6  7  8 NA NA

x4 <- remove_user_na(x1, user_na_to_na = TRUE)
x4

#> <Labelled integer>
#>  [1]  1  2  3  4  5  6  7  8 NA NA
#> 
#> Labels:
#>  value label
#>      1  Good
#>      8   Bad