remove_labels {labelled}	R Documentation

Remove variable label, value labels and user defined missing values

Description

Use remove_var_label to remove variable label, remove_val_labels to remove value labels, remove_user_na to remove user defined missing values (na_values and na_range) and remove_labels to remove all.

Usage

remove_labels(x, user_na_to_na = FALSE)

remove_var_label(x)

remove_val_labels(x)

remove_user_na(x, user_na_to_na = FALSE)

Arguments

x
A vector or a data frame.

user_na_to_na
Convert user defined missing values into NA?

Details

Be careful with remove_user_na and remove_labels, user defined missing values will not be automatically converted to NA, except if you specify user_na_to_na = TRUE. user_na_to_na(x) is an equivalent of remove_user_na(x, user_na_to_na = TRUE).

Examples

x1 <- labelled_spss(1:10, c(Good = 1, Bad = 8), na_values = c(9, 10))
var_label(x1) <- "A variable"
x1

x2 <- remove_labels(x1)
x2
x3 <- remove_labels(x1, user_na_to_na = TRUE)
x3
x4 <- remove_user_na(x1, user_na_to_na = TRUE)
x4

[Package labelled version 2.0.2 Index]